backend/requirements.txt
Django~=5.2.4
djangorestframework~=3.16.0
django-cors-headers
psycopg2-binary
django_filter
pytest
pytest-django
pytest-cov
factory-boy
model-bakery
locust